Revert node black/whitelist feature
Node whitelisting to create a heterogenous system of sworms is not the
right approach. The correct approach would be to only start the sworms
on the nodes that need them, based on the OTP applications that run on
those nodes.

Note that the current manager does not currently support this and
assumes that all nodes run all sworms.