Use {:shutdown, term} to communicate worker pid startup conflict
Although maybe misusing the `:shutdown` return value here, this seems
like a better way to communicate the conflicting worker PID to the
starter call site.

Previously, :ignore was returned from the delegate and then a lookup
was done, which could lead to a race condition with the registration
CRD update.